// src/navigator/types.ts
export enum AppRoutes {
  APP = 'App',
  LOGIN = 'Login',
  LOGOUT = 'Logout',
  REGISTER = 'Register',
  OTP = 'OTP',
  UPDATE_PROFILE = 'UpdateProfile',
  PROFILE = 'Profile',
  HOME = 'Home',
  SETTING = 'Setting',
  FORGOT_PASSWORD = 'ForgotPassword',
  RESET_PASSWORD = 'ResetPassword',
  CHANGE_PASSWORD = 'ChangePassword',
  CODE_CONFIRMATION = 'CodeConfirmation',
  UPDATE_PHONE = 'UpdatePhone',
  CREATE_INVOICE = 'CreateInvoice',
  HOADON_GTGTMTT = 'HoaDonGTGTMTT',
  KY_HIEU_MAUHOADON = 'KyHieuMauHoaDon',
  DS_KHACHHANG_DOITAC = 'DSkhachHangDoiTac',
  DS_HANGHOA_DICHVU = 'DShangHoaDichVu',
  THONGTIN_NGUOIMUAHANG = 'ThongTinNguoiMuaHang',
  THONGBAO_SAISOT = 'thongBaoSaiSot',
  XULY_XOABOB1 = 'XuLyXoaBoB1',
  XULY_XOABOB2 = 'XuLyXoaBoB2',
  DS_HOADON_SAISOT = 'dsHoaDonSaiSot',
  DS_THONGBAO_SAISOT = 'dsThongBaoSaiSot',
  LAPTHONGBAO_SAISOT = 'lapThongBaoSaiSot',
  ACCOUNR_SCREEN = 'accountSceen',
  MAIL_LIST_SCREEN = 'mailListScreen',
  INVOICE_HELP_SCREEN = 'invoiceHelpScreen',
  INVOICE_SETTING = 'invoiceSetting',
  CHANGE_PIN = 'changePIN',
  LAUNCH = 'launch',
  INVOICE_CUSMER = 'invoiceCusmer',
  GOODS_SCREEN = 'goodsScreen',
  DONVI_TINH = 'donViTinh',
  THONGTINH_DONVI_TINH = 'thongTinDonViTinh',
  THEMSUA_DONVI_TINH = 'themSuaDonViTinh',
  MAIN_TAB = 'MainTab',
  HINHTHUC_THANHTOAN = 'hinhThucThanhToan',
  THONGTIN_HINHTHUC_THANHTOAN = 'thongTinHinhThucThanhToan',
  THEMSUA_HINHTHUC_THANHTOAN = 'themSuaHinhThucThanhToan',
  TIEN_THANHTOAN = 'tienThanhToan',
  THONGTIN_TIEN_THANHTOAN = 'thongTinTienThanhToan',
  THEMSUA_TIEN_THANHTOAN = 'themSuaTienThanhToan',
  LOAI_HANGHOA = 'loaiHangHoa',
  THONGTIN_LOAI_HANGHOA = 'thongTinLoaiHangHoa',
  THEMSUA_LOAI_HANGHOA = 'themSuaLoaiHangHoa',
  THONGDIEP_GUI = 'thongDiepGui',
  THONGDIEP_NHANH = 'thongDiepNhan',
  SELECT_FORM = 'selectForm',
  RENAME_FORM = 'renameForm',
  EDIT_FORM = 'editForm',
  INFO_FORM = 'inforForm',
  STEPS = 'steps',
  RELEASE_ENVOICE = 'releaseEnvoice',
  START_RELEASE_ENVOICE = 'startReleaseEnvoice',
  SUBBMIT_AGENCY = 'submitAgency',
  START_USE = 'startUse',
  ARROUND_NUMBER = 'aroundNumber',
  FILTER_INVOICE = 'filterInvoice',
  PRODUCT_EDIT = 'productEdit',
  ADD_GOODS = 'addGoods',
  ADD_CUSTOMER = 'addCustomer',
  DETAIL_CUSTOMER = 'detailCustomer',
  DETAIL_GOODS = 'detailGoods',
  XACNHAN_MAPIN = 'xacNhanMaPin',
  DIALOG = 'dialog',
  DIALLOG_SUCCESS = 'dialogSuccess',
  ACTION_SHEET = 'actionSheet',
  CONFIRM_PASS = 'confirmPass',
  DIALOG_LOADING = 'dialogLoading',
  ACTION_DRAWER = 'actionDrawer',
  PRODUCT_LIST = 'productList',
  UNIT_LIST = 'unitList',
  CUSTOMER_LIST = 'customerList',
  LOAI_HANG = 'loaiHang',
  DRAWER_GOODS = 'drawerGoods',
  ACTION_PREVIEW = 'actionPreview',
  XEMTHONGBAO_PREVIEW = 'xemThongBaoPreview',
  FILTER_REPORT = 'filterReport',
  FILTER_SENDINVOICE_CQT = 'filterSendInvoiceCQT',
  FILTER_MOVEINVOICECQT = 'filterMoveInvoiceCQT',
  REPORT_DETAIL = 'reportDetail',
  SENT_INFOR_CQT_POPUP = 'sentInfoCQTpopup',
  SEND_INVOICE_MAIL = 'sendInvoiceMail',
  PRPCESS_ENIVOICE = 'processEnivoice',
  DRAWER = 'drawer'
 
}

export enum AppTabs {
  HOME_TAB = 'HomeTab',
  HomeScreen = 'HomeScreen',
  listInvoice='listInvoice',
  list = 'list',
  Statistic = 'Statistic',
  DashboardSceen = 'DashboardSceen',
  Report ="Report",
  reportScreen='reportScreen',
  sendCQT='sendCQT',
}